warning fix

+++ Thomas/src/Platform/OpenGL/OpenGLContext.cpp
@@ -72,9 +72,9 @@ namespace Thomas
 			TH_CORE_INFO("Current OpenGL Context is NOT double buffered");
 		}
 		
-		TH_CORE_INFO("Maximum Vertex Count: {0}", (int)int_4);
-		TH_CORE_INFO("Maximum Indices Count: {0}", (int)int_5);
-		TH_CORE_INFO("GL Maximum texture size: {0}", (int)int_6);
-		TH_CORE_INFO("Maximum Viewport Dimensions: {0}", (int)int_7);
+		TH_CORE_INFO("Maximum Vertex Count: {0}", (unsigned long long)int_4);
+		TH_CORE_INFO("Maximum Indices Count: {0}", (unsigned long long)int_5);
+		TH_CORE_INFO("GL Maximum texture size: {0}", (unsigned long long)int_6);
+		TH_CORE_INFO("Maximum Viewport Dimensions: {0}", (unsigned long long)int_7);
 	}
 }
